+++ Client/package.json
@@ -11,7 +11,7 @@
   },
   "dependencies": {
     "@mantine/core": "^7.16.3",
-    "@radix-ui/react-slot": "^1.1.1",
+    "@radix-ui/react-slot": "^1.1.2",
     "@tanstack/react-query": "^5.64.2",
     "animate.css": "^4.1.1",
     "autoprefixer": "^10.4.18",
