+++ Client/src/components/ui/button.jsx
@@ -34,10 +34,6 @@ const buttonVariants = cva(
   }
 )
 
-
-// Sadui ui Add By Manish Prajapati
-
-
 const Button = React.forwardRef(({ className, variant, size, asChild = false, ...props }, ref) => {
   const Comp = asChild ? Slot : "button"
   return (
