+++ Client/src/pages/ProjectInfo.jsx
@@ -1,5 +1,4 @@
 import { useEffect, useState } from "react";
-import axios from "axios";
 import {
   Github,
   Users,
