+++ Client/src/pages/home/goalsAndNotes/NotesComponent.jsx
